pyproject.toml
[tool.poetry]
name = "ml-api"
version = "0.1.0"
description = ""
authors = ["dumaas <normify@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
psycopg2-binary = "^2.8.6"
numpy = "^1.20.3"
pandas = "^1.2.4"
sklearn = "^0.0"
joblib = "^1.0.1"
nltk = "^3.6.2"
surprise = "^0.1"
jupyter = "^1.0.0"
ipykernel = "^5.5.5"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
gunicorn = "^20.1.0"
django-allauth = "^0.44.0"
dj-rest-auth = "^2.1.5"
django-cors-headers = "^3.7.0"
djangorestframework-simplejwt = "^4.7.0"
drf-yasg = "^1.20.0"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
